==> 2. feature engineering.ipynb <==
"""
Feature engineering 

As from last notebook, we need to do following in feature engineering
- exclude 0 sales
- exclude store closure
- state holiday: one hot encoding
- date: add year, month, week, quarter
"""

import pandas as pd
import warnings
import seaborn as sns

warnings.simplefilter("ignore")
pd.options.display.max_columns = None

input_path = 'rossmann/'
input_file = 'train.csv'
processed_path = 'processed/'
processed_file = 'processed_data.csv'


df_train = pd.read_csv(input_path + input_file)

# Exclude 0 sales and store closure

def exclude_store_closure_and_zero(df):
    df = df.loc[df['Sales']>0]
    return df.loc[df['Open']==1].reset_index(drop=True)

df = exclude_store_closure_and_zero(df_train)

df.shape

df

# Deal with state holiday

def one_hot_encoding_state_holiday(df):
    """a = public holiday, b = Easter holiday, c = Christmas, 0 = None
    """
    df.loc[df['StateHoliday'] == 0, 'StateHoliday'] = '0'
    state_dummy = df.pop('StateHoliday')
    state_dummy = pd.get_dummies(state_dummy)
    state_dummy.rename(columns={'0': 'holiday_no', 'a': 'holiday_public', 
                               'b': 'holiday_easter', 'c':'holiday_christmas'}, inplace=True)
    
    combined = pd.concat([df, state_dummy], axis =1)
    return combined

df = one_hot_encoding_state_holiday(df)

# Date

def add_year_month(df):
    df['Date'] = pd.to_datetime(df['Date'], infer_datetime_format=True)
    df = df.sort_values('Date', ascending = True)
    df['year'] = df['Date'].dt.year
#     df['month'] = df['Date'].dt.month
#     df['quarter'] = df['Date'].dt.quarter
    df['weekofyear'] = df['Date'].dt.weekofyear
    return df

df = add_year_month(df)

df

# Combine all together 

def feature_engineering(df):
    df = exclude_store_closure_and_zero(df)
    df = one_hot_encoding_state_holiday(df)
    df = add_year_month(df)
    return df

df = feature_engineering(df_train)

df

# Save to processed folder

df.to_csv(processed_path + processed_file, index=False)
