==> 3. Regressions.ipynb <==
# Regression 

This notebook is mainly for basic regressions and evaluations

import pandas as pd
import numpy as np

import warnings
import pickle
warnings.simplefilter("ignore")
pd.options.display.max_columns = None

from sklearn import linear_model
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import StandardScaler
from sklearn.svm import SVR
from sklearn import neighbors
import xgboost as xgb
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import Normalizer

from keras.models import Sequential
from keras.models import Model as KerasModel
from keras.layers import Input, Dense, Activation, Reshape, Concatenate
from keras.callbacks import ModelCheckpoint



TRAIN_RATIO = 0.9

SHUFFLE_DATA = True
ONE_HOT_AS_INPUT = False # in addition to state holidays, use one hot for other categories

xgb_model_dump = 'models/xgb_model.pickle'
rf_model_dump = 'models/rf_model.pickle'
nn_model_dump = 'models/nn_model.pickle'

# Read training data

processed_path = 'processed/'
processed_file = 'processed_data.csv'
X = pd.read_csv(processed_path + processed_file)

X.columns

regression_cols = ['Store', 'DayOfWeek', 'Sales', 'Promo',
       'SchoolHoliday', 'holiday_no', 'holiday_public', 'holiday_easter',
       'holiday_christmas', 'year', 'weekofyear']

X = X[regression_cols]

y = X.pop('Sales')

if ONE_HOT_AS_INPUT:
    print("Using one-hot encoding as input for model")
    cols_for_one_hot = ['Store', 'DayOfWeek', 'weekofyear']
    one_hot_df = X[cols_for_one_hot]
    X = X.drop(columns = cols_for_one_hot)
    
    enc = OneHotEncoder(sparse=False)
    enc.fit(one_hot_df)
    one_hot_df = enc.transform(one_hot_df)
    X = X.to_numpy()
    X = pd.concat([X, one_hot_df], axis=1)

# Use last 10% data as validation data
num_record = len(X)
train_size = int(TRAIN_RATIO * num_record)
X_train = X[:train_size]
y_train = y[:train_size]
X_val = X[train_size:]
y_val = y[train_size:]

if SHUFFLE_DATA:
    print("Using shuffled data")
    sh = np.arange(X_train.shape[0])
    np.random.shuffle(sh)
    X_train = X_train.iloc[sh]
    y_train = y_train.iloc[sh]

# Build models

class Model(object):
    def evaluate(self, X_val, y_val):
        assert (min(y_val) > 0)
        guessed_sales = self.guess(X_val)
        relative_err = np.absolute((y_val - guessed_sales) / y_val)
        result = np.sum(relative_err) / len(y_val)
        return result

    def rsqr_score(self, X_val, y_val):
        assert (min(y_val) > 0)
        guessed_sales = self.guess(X_val)
        SS_Residual = sum((y_val-guessed_sales)**2)
        SS_Total = sum((y_val-np.mean(y_val))**2)
        r_squared = 1 - (float(SS_Residual))/SS_Total
        adjusted_r_squared = 1 - (1-r_squared)*(len(y_val)-1)/(len(y_val)-X_val.shape[1]-1)
        return adjusted_r_squared
    

class LinearModel(Model):
    def __init__(self, X_train, y_train, X_val, y_val):
        super().__init__()
        self.model = linear_model.LinearRegression()
        self.model.fit(X_train, np.log(y_train))
        print("Result on validation set is: ", self.evaluate(X_val, y_val))
        print("R squared on validation set is: ", self.rsqr_score(X_val, y_val))

    def guess(self, features):
        return np.exp(self.model.predict(features))

    def model(self):
        return self.model

class RidgeRegression(Model):
    def __init__(self, X_train, y_train, X_val, y_val):
        super().__init__()
        self.model = linear_model.Ridge(alpha=1.0)
        self.model.fit(X_train, np.log(y_train))
        print("MAPE on validation set is: ", self.evaluate(X_val, y_val))
        print("R squared on validation set is: ", self.rsqr_score(X_val, y_val))

    def guess(self, features):
        return np.exp(self.model.predict(features))
    
    def model(self):
        return self.model

class RF(Model):
    def __init__(self, X_train, y_train, X_val, y_val):
        super().__init__()
        self.model = RandomForestRegressor(n_estimators=200,
                                         verbose=True,
                                         max_depth=35,
                                         min_samples_split=2,
                                         min_samples_leaf=1
                                         )
        self.model.fit(X_train, np.log(y_train))
        print("MAPE on validation set is: ", self.evaluate(X_val, y_val))
        print("R squared on validation set is: ", self.rsqr_score(X_val, y_val))

    def guess(self, features):
        return np.exp(self.model.predict(features))

    def model(self):
        return self.model


class SVM(Model):

    def __init__(self, X_train, y_train, X_val, y_val):
        super().__init__()
        self.X_train = X_train
        self.y_train = y_train
        self.__normalize_data()
        self.model = SVR(kernel='linear', degree=3, gamma='auto', coef0=0.0, tol=0.001,
                       C=1.0, epsilon=0.1, shrinking=True, cache_size=200, verbose=False, max_iter=-1)

        self.model.fit(self.X_train, np.log(self.y_train))
        print("MAPE on validation set is: ", self.evaluate(X_val, y_val))
        print("R squared on validation set is: ", self.rsqr_score(X_val, y_val))

    def __normalize_data(self):
        self.scaler = StandardScaler()
        self.X_train = self.scaler.fit_transform(self.X_train)

    def guess(self, feature):
        return np.exp(self.model.predict(feature))

class XGB(Model):
    def __init__(self, X_train, y_train, X_val, y_val):
        super().__init__()
        dtrain = xgb.DMatrix(X_train, label=np.log(y_train))
        evallist = [(dtrain, 'train')]
        params = {'nthread': -1,
                  'max_depth': 7,
                  'eta': 0.02,
                  'silent': 1,
                  'objective': 'reg:squarederror',
                  'colsample_bytree': 0.7,
                  'subsample': 0.7}
        num_round = 3000
        self.model = xgb.train(params,
                             dtrain,
                             num_round,
                             evallist,
                             verbose_eval = 50,
                             early_stopping_rounds = 50)
        print("MAPE on validation set is: ", self.evaluate(X_val, y_val))
        print("R squared on validation set is: ", self.rsqr_score(X_val, y_val))

    def guess(self, features):
        dtest = xgb.DMatrix(features)
        return np.exp(self.model.predict(dtest))

class KNN(Model):

    def __init__(self, X_train, y_train, X_val, y_val):
        super().__init__()
        self.normalizer = Normalizer()
        self.normalizer.fit(X_train)
        self.model = neighbors.KNeighborsRegressor(n_neighbors=10, weights='distance', p=1)
        self.model.fit(self.normalizer.transform(X_train), np.log(y_train))
        print("MAPE on validation set is: ", self.evaluate(self.normalizer.transform(X_val), y_val))
        print("R squared on validation set is: ", self.rsqr_score(self.normalizer.transform(X_val), y_val))

    def guess(self, feature):
        return np.exp(self.model.predict(self.normalizer.transform(feature)))

class NN(Model):

    def __init__(self, X_train, y_train, X_val, y_val):
        super().__init__()
        self.epochs = 10
        self.checkpointer = ModelCheckpoint(filepath="models/best_model_weights.hdf5", verbose=1, save_best_only=True)
        self.max_log_y = max(np.max(np.log(y_train)), np.max(np.log(y_val)))
        self.__build_keras_model()
        self.fit(X_train, y_train, X_val, y_val)

    def __build_keras_model(self):
        self.model = Sequential()
        self.model.add(Dense(1000, kernel_initializer="uniform", input_dim=1181))
        self.model.add(Activation('relu'))
        self.model.add(Dense(500, kernel_initializer="uniform"))
        self.model.add(Activation('relu'))
        self.model.add(Dense(1))
        self.model.add(Activation('sigmoid'))

        self.model.compile(loss='mean_absolute_error', optimizer='adam')

    def _val_for_fit(self, val):
        val = np.log(val) / self.max_log_y
        return val

    def _val_for_pred(self, val):
        return np.exp(val * self.max_log_y)

    def fit(self, X_train, y_train, X_val, y_val):
        self.model.fit(X_train, self._val_for_fit(y_train),
                       validation_data=(X_val, self._val_for_fit(y_val)),
                       epochs=self.epochs, batch_size=128,
                       # callbacks=[self.checkpointer],
                       )
        # self.model.load_weights('models/best_model_weights.hdf5')
        print("MAPE on validation set is: ", self.evaluate(X_val, y_val))
        print("R squared on validation set is: ", self.rsqr_score(X_val, y_val))

    def guess(self, features):
        result = self.model.predict(features).flatten()
        return self._val_for_pred(result)


# Training

print("Fitting Linear Model...")
Linear_model = LinearModel(X_train, y_train, X_val, y_val)

print("Fitting Ridge Model...")
ridge_model = RidgeRegression(X_train, y_train, X_val, y_val)

print("Fitting KNN...")
KNN(X_train, y_train, X_val, y_val)

print("Fitting SVM...")
SVM(X_train, y_train, X_val, y_val)

print("Fitting Random Forest Model...")
rf_model = RF(X_train, y_train, X_val, y_val)

print("Fitting XGB Model...")
xgb_model = XGB(X_train, y_train, X_val, y_val)

print("Fitting NN...")
nn_model = NN(X_train, y_train, X_val, y_val)

# Save models

with open(rf_model_dump, 'wb') as f:
    pickle.dump(rf_model, f, -1)

with open(xgb_model_dump, 'wb') as f:
    pickle.dump(xgb_model, f, -1)

with open(nn_model_dump, 'wb') as f:
    pickle.dump(nn_model, f, -1)
