==> 9. pull forward.ipynb <==
Pull forward is evaluated at customer item/category level

The idea is that certain customers stock up things on promotion, so the buying pattern changes because of promotions. 

There are many limitations regarding pull forward. 

The major one is that not every category are eligible for pull forward analysis, example: computer, calendar.

Simply nobody is going to stock up these products becasue they are on promotion.

The most likely categories are the ones that customers purchase frequently, such as toilet paper. 

There are 2 different ways to calculate pull forward effect, but the main idea is the same. Differentiate the purchase pattern by promotion, if certain customer groups behave totally different becasue of promotion, then we say that the promotion has pull forward impact. 
    1. calculate the buying frequencies with and without promotions. If the promotion does not have any effect, the distribution should be similar. Z-test is a good way to understand the behavior of the 2 distributions. if Z score is largert than 3, then we can say that the promotion has significant impact on consumer behavior. 
    2. use regression model (poisson) to calculate the number of transactions that a customer will make within a certain period of time (eg future 90 days)
